Moved profileSystemCodes out of data directory
